requirements.txt
# external requirements
importlib-metadata<5.0
flake8
python-dotenv>=0.5.1
pytest
numpy
pandas>=1.3.0
seaborn
matplotlib
scipy
hydra-core
hydra_colorlog
mlflow
scikit-learn
tables
ray[tune]
torchcde
hawkes
pytorch-lightning==1.4.5
torchmetrics==0.6.0
